+++ pyveoliaidf/enum.py
@@ -1,8 +1,9 @@
 from enum import Enum
 
 class PropertyNameEnum(Enum):
-    TIME = 'time'
-    TOTAL_LITER = 'total_liter'
-    DAILY_LITER = 'daily_liter'
-    TYPE = 'type'
+    TIME = "time"
+    TOTAL_LITER = "total_liter"
+    DAILY_LITER = "daily_liter"
+    TYPE = "type"
+    TIMESTAMP = "timestamp"