+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Use PropertyNameEnum type to store all property names.
 - Add LoginError exception raised when PyVeoliaIDF is unable to sign in the Veolia Web site with the given username/password.
+- Add timestamp property that contains date/time when the data has been retrieved.
 
 [Unreleased]: https://github.com/ssenart/PyGazpar/compare/0.1.8...HEAD
 [0.1.8]: https://github.com/ssenart/PyGazpar/compare/0.1.7...0.1.8